+++ .env.example
@@ -1,0 +1,5 @@
+NODE_ENV="production"
+VITE_API_URL="http://localhost:3000/"
+DATABASE_URL="file:./dev.db"
+DISCORD_NOTIFICATION_WEBHOOK="https://discord.com/api/webhooks/..."
+TIMEZONE="America/New_York"
