--- api/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "api",
-  "version": "1.0.0",
-  "main": "dist/src/index.js",
-  "scripts": {
-    "start": "tsc && concurrently \"node dist/index.js\" \"node dist/services/Tasks.js\"",
-    "dev:api": "ts-node ./src/index.ts",
-    "dev:cron": "ts-node ./src/services/Tasks.ts",
-    "db:push": "npx prisma db push",
-    "db:pull": "npx prisma db pull"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "devDependencies": {
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
-    "@types/node-cron": "^3.0.11",
-    "concurrently": "^8.2.2",
-    "nodemon": "^3.1.4",
-    "prisma": "^5.17.0",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.5.4"
-  },
-  "dependencies": {
-    "@prisma/client": "^5.17.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "moment": "^2.30.1",
-    "node-cron": "^3.0.3",
-    "zod": "^3.23.8"
-  }
-}
