--- api/src/routers/index.ts
@@ -1,7 +1,0 @@
-import Budget from "./Budget"
-import Analytics from './Analytics';
-
-export default {
-    Budget,
-    Analytics
-}
