+++ package.json
@@ -1,0 +1,49 @@
+{
+  "name": "web",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "web:dev": "vite",
+    "web:dev:host": "vite --host",
+    "web:build": "tsc -b && vite build",
+    "api:dev": "tsc --project ./src/router && node ./src/router/dist",
+    "api:dev:cron": "tsc --project ./src/router && node ./src/router/dist/services/Tasks.js",
+    "api:db:push": "npx prisma db push",
+    "api:db:pull": "npx prisma db pull",
+    "start": "tsc --project ./src/router && npm run web:build && concurrently \"node ./src/router/dist\" \"node ./src/router/dist/services/Tasks.js\""
+  },
+  "dependencies": {
+    "@prisma/client": "^5.17.0",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "http-proxy-middleware": "^3.0.2",
+    "moment": "^2.30.1",
+    "node-cron": "^3.0.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "zod": "^3.23.8"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/node-cron": "^3.0.11",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^7.15.0",
+    "@typescript-eslint/parser": "^7.15.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.19",
+    "concurrently": "^8.2.2",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
+    "nodemon": "^3.1.4",
+    "postcss": "^8.4.40",
+    "prisma": "^5.17.0",
+    "tailwindcss": "^3.4.7",
+    "typescript": "^5.5.4",
+    "vite": "^5.3.4"
+  }
+}
