+++ src/components/Icons/Trash.tsx
