+++ src/components/ItemDisplay.tsx
