+++ src/router/routers/index.ts
@@ -1,0 +1,7 @@
+import Budget from "./Budget.js"
+import Analytics from './Analytics.js';
+
+export default {
+    Budget,
+    Analytics
+}
