+++ src/router/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
     "target": "ES2020",
-    "module": "commonjs",
-    "rootDir": "./src",
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
+    "rootDir": "./",
     "outDir": "./dist",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
@@ -10,10 +11,7 @@
     "noImplicitAny": true,
     "skipLibCheck": true
   },
-  "exclude": [
-    "node_modules"
-  ],
   "include": [
-    "./src/**/*.ts"
+    "./**/*.ts"
   ]
 }
