+++ tsconfig.app.json
@@ -23,5 +23,5 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"]
+  "exclude": ["./src/router", "node_modules"]
 }
